Done reservations

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, abort, request, redirect, url_for, flash
+from flask import Flask, render_template, abort, request, redirect, url_for, session
 from flask_mysqldb import MySQL
 
 app = Flask(__name__)
@@ -40,6 +40,8 @@ def single_book(book_id):
 
 @app.route("/books/add", methods=['GET', 'POST'])
 def add_book():
+    if session.get('user_type') != 2:
+        return redirect("/books")
     if request.method == 'POST':
         title = request.form['title']
         genre = request.form['genre']
@@ -57,6 +59,8 @@ def add_book():
 
 @app.route("/books/edit/<int:book_id>", methods=['GET', 'POST'])
 def update_book(book_id):
+    if session.get('user_type') != 2:
+        return redirect("/books")
     if request.method == 'POST':
         title = request.form['title']
         genre = request.form['genre']
@@ -79,6 +83,8 @@ def update_book(book_id):
 
 @app.route("/books/delete/<int:book_id>")
 def delete_book(book_id):
+    if session.get('user_type') != 2:
+        return redirect("/books")
     try:
         cursor = db.connection.cursor()
         cursor.execute(f"DELETE FROM knjiga WHERE id={book_id}")
@@ -87,6 +93,55 @@ def delete_book(book_id):
         return redirect("/books")
     except:
         abort(404)
+
+@app.route("/reservations")
+def all_reservations():
+    if not session.get('is_logged'):
+        return redirect("/login")
+    if session['user_type'] == 1:
+        cursor = db.connection.cursor()
+        cursor.execute(f"SELECT rezervacije.id, ime, prezime, naslov, datum_preuzimanja, datum_vracanja FROM korisnik JOIN rezervacije ON rezervacije.korisnik_id = korisnik.id JOIN knjiga ON rezervacije.knjiga_id = knjiga.id WHERE korisnik.id={session['id']}")
+        reservations = cursor.fetchall()
+        cursor.close()
+    else:
+        cursor = db.connection.cursor()
+        cursor.execute(f"SELECT rezervacije.id, ime, prezime, naslov, datum_preuzimanja, datum_vracanja FROM korisnik JOIN rezervacije ON rezervacije.korisnik_id = korisnik.id JOIN knjiga ON rezervacije.knjiga_id = knjiga.id")
+        reservations = cursor.fetchall()
+        cursor.close()
+    return render_template("reservations.html",reservations=reservations)
+
+@app.route("/reservations/add",  methods=['GET', 'POST'])
+def add_reservation():
+    if not session.get('is_logged'):
+        return redirect("/login")
+    if request.method == 'POST':
+        book = request.form['book']
+        user = session['id']
+        issue_start = request.form['issue_start']
+        issue_end = request.form['issue_end']
+        cursor = db.connection.cursor()
+        cursor.execute(f"INSERT INTO rezervacije VALUES (NULL,{book},{user},'{issue_start}','{issue_end}')")
+        db.connection.commit()
+        cursor.close()
+        return redirect('/reservations')
+    cursor = db.connection.cursor()
+    cursor.execute(f"SELECT id, naslov FROM knjiga ORDER BY naslov")
+    books = cursor.fetchall()
+    return render_template("reservations_form.html", books=books)
+
+@app.route("/reservations/delete/<int:id>")
+def delete_reservation(id):
+    if session.get('user_type') != 2:
+        return redirect("/reservations")
+    try:
+        cursor = db.connection.cursor()
+        cursor.execute(f"DELETE FROM rezervacije WHERE id={id}")
+        db.connection.commit()
+        cursor.close()
+        return redirect("/reservations")
+    except:
+        abort(404)
+    
 
 import auth
